consolidate trading logic
